Item hover background color change
Fixed all the issues I was having with the sprite earlier, looks good even on the category items page. Note that the shadow image has been updated (I left the original image in the same directory still I just renamed it "grid-item-shadow2.png")

I'm still not entirely sure about the rollover color but yea that is an easy change.

Signed-off-by: JJ <JJ@onlyhuman.me>